EDA on Client Loan Repayment
This case study aims at applying EDA in a real business scenario. Apart from applying the techniques of EDA, this case study aims at developing a basic understanding of risk analytics in banking and financial services and understand how data is used to minimise the risk of losing money while lending to customers.